pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scholarboard-ai"
version = "0.1.0"
description = "Analyze and visualize scholars based on their research areas"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "User", email = "user@example.com"}
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openai>=1.65.3",
    "numpy>=1.24.4",
    "plotly>=6.0.0",
    "pandas>=2.0.3",
    "numba>=0.58.1",
    "llvmlite>=0.41.1",
    "umap-learn>=0.5.7",
    "matplotlib>=3.7.5",
    "xarray>=2023.1.0",
    "netCDF4>=1.6.0",
    "scikit-learn>=1.2.0",
]

[project.scripts]
scholarboard = "scholar_board.cli:main"

[tool.setuptools]
packages = ["scholar_board"]

[tool.uv]
dev-dependencies = [] 
